couldiat_project/requirements.txt
asgiref==3.10.0
asttokens==3.0.0
boto3==1.40.55
botocore==1.40.55
decorator==5.2.1
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-extensions==4.1
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
executing==2.2.1
gunicorn==23.0.0
inflection==0.5.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
whitenoise==6.11.0
